pyproject.toml
[project]
name = "aipaca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jorge-Alda", email = "jorge.alda@pd.infn.it" }
]
requires-python = ">=3.13"
dependencies = [
    "alpaca-alps==1.2.0rc2",
    "fastmcp>=3.2.4",
    "pydantic>=2.13.3",
]

[project.scripts]
aipaca-mcp = "aipaca.mcp_server:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
